+++ packages/scene-sdk/src/index.ts
@@ -2,3 +2,5 @@ export * from "./compatibility.js";
 export * from "./behaviorScheduler.js";
 export * from "./commandValidation.js";
 export * from "./protocol.js";
+
+export * from "./scriptModuleAdapter.js";